requirements.txt
m3u8==0.8.0
requests
pycryptodome
